backend/requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
pydantic==2.7.3
starlette==0.37.2
python-dotenv==1.0.1
python-multipart==0.0.9
openai==0.28
websockets==11.0.3
python-jose==3.3.0
